[Meta\Docs] move logic to helper function
